- Cleanup various warnings     - WINSCR: Implement new configuration dialog with new parameters     - WINSCR: Remove crufty code     - WINSCR: Update copyrights in resource files
    client/
        app.cpp
    clientscr/
        boinc_ss.h
        boinc_ss.rc
        boinc_ss_opengl.h
        boinc_ss_opengl.rc
    clientscr/
        screensaver.cpp, .h
        screensaver_win.cpp, .h
    lib/
        common_defs.h
        crypt.cpp
        win_util.cpp
    samples/worker/
        worker.cpp
    win_build/
        boinc_ss.vcproj
        multi_thread.vcproj
        sleeper.vcproj
        ss_app.vcproj
        uc2_graphics.vcproj

svn path=/branches/boinc_core_release_6_10/; revision=19942

+++ lib/crypt.cpp
@@ -589,7 +589,7 @@ char *check_validity(
     FILE* of = boinc_fopen(origFile, "r");
     if (!of) return NULL;
     MD5_Init(&md5CTX);
-    while (0 != (rbytes = fread(rbuf, 1, sizeof(rbuf), of))) {
+    while (0 != (rbytes = (int)fread(rbuf, 1, sizeof(rbuf), of))) {
 	    MD5_Update(&md5CTX, rbuf, rbytes);
     }
     MD5_Final(md5_md, &md5CTX);
@@ -637,7 +637,7 @@ int cert_verify_file(
     FILE* of = boinc_fopen(origFile, "r");
     if (!of) return false;
     MD5_Init(&md5CTX);
-    while (0 != (rbytes = fread(rbuf, 1, sizeof(rbuf), of))) {
+    while (0 != (rbytes = (int)fread(rbuf, 1, sizeof(rbuf), of))) {
 	    MD5_Update(&md5CTX, rbuf, rbytes);
     }
     MD5_Final(md5_md, &md5CTX);
